requirements.txt
google-adk>=1.13.0  
mem0ai>=0.1.0  
python-dotenv>=1.0.0  
pydantic>=2.0.0